+++ framework/Utils/ClassValueCacheUtils.php
@@ -11,12 +11,18 @@ namespace Rid\Utils;
 
 trait ClassValueCacheUtils
 {
+    protected static function getStaticCacheNameSpace(): string
+    {
+        return 'Cache:default_static';
+    }
+
     protected function getCacheNameSpace(): string
     {
         return 'Cache:default';
     }
 
-    protected function getCacheValue($key, $closure)
+    // Get from redis cache then database
+    final protected function getCacheValue($key, $closure)
     {
         if (is_null($this->$key)) {
             $this->$key = app()->redis->hGet($this->getCacheNameSpace(), $key);
@@ -26,5 +32,15 @@ trait ClassValueCacheUtils
             }
         }
         return $this->$key;
+    }
+
+    final protected static function getStaticCacheValue($key, $closure, $ttl = 86400)
+    {
+        $value = app()->redis->get(static::getStaticCacheNameSpace() . ':' . $key);   // FIXME use zset
+        if (false === $value) {
+            $value = $closure();
+            app()->redis->set(static::getStaticCacheNameSpace() . ':' . $key, $value, $ttl);
+        }
+        return $value;
     }
 }